cleanup mdx dic
